backend/pyproject.toml
[tool.poetry]
name = "AI-Powered-CRO-AUDIT"
version = "0.1.0"
description = ""
authors = ["Hamza-Ali <hamzaaly105@gmail.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.scripts]
start = "app.cli:start"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.1"
uvicorn = "0.24.0"
python-dotenv = "1.0.0"
pymongo = "4.3.3"
motor = "3.1.1"
pydantic = "2.11.7"
pydantic-settings = "2.10.1"
email-validator = "2.2.0"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-multipart = "0.0.6"
sentry-sdk = "1.32.0"
python-json-logger = "^3.2.1"
sendgrid = "^6.11.0"
pyjwt = "^2.10.1"
poetry-dotenv-plugin = "^0.2.0"
poetry = "^2.0.0"
sqlalchemy = "^2.0.41"
sqlmodel = "^0.0.24"
asyncpg = "^0.30.0"
nest-asyncio = "^1.6.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.14.0"
pre-commit = "^4.0.1"
autoflake = "^2.2.0"
bandit = "^1.7.5"
types-python-jose = "^3.3.4"
types-passlib = "^1.7.7"
alembic = "^1.16.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs  # exclude docs directory
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
src_paths = ["app"]
skip = ["__init__.py"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".git", "__pycache__"]
targets = ["app"]

[tool.autoflake]
recursive = true
in-place = true
remove-all-unused-imports = true
remove-unused-variables = true
exclude = ["__init__.py", "migrations"]
